Pluggable cash selection using H2 as default implementation.
